[improvement] Exception handling

+++ src/dev/norbu/donezo/cli/command/Delete.java
@@ -1,10 +1,9 @@
 package dev.norbu.donezo.cli.command;
 
+import dev.norbu.donezo.cli.exception.InvalidInputException;
 import dev.norbu.donezo.service.TaskService;
 
 import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Objects;
 
 public class Delete
         implements Command {
@@ -17,15 +16,12 @@ public class Delete
 
     @Override
     public void execute(final List<String> args) {
-        Objects.requireNonNull(args);
-        try {
-            final String id = args.getFirst();
-            taskService.deleteById(id);
-        } catch (IllegalArgumentException _) {
-            System.err.printf("Invalid task ID: %s", args.getFirst());
-        } catch (NoSuchElementException _) {
-            System.err.printf("No task with ID: %s found.", args.getFirst());
+        if (args.isEmpty()) {
+            throw new InvalidInputException("No task ID provided.");
         }
+        final String id = args.getFirst();
+        taskService.deleteById(id);
+        System.out.printf("Task '%s' deleted.", id);
     }
 
     @Override
